pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "novel-rewriter"
version = "0.1.0"
description = "仿写百万字小说软件 - 基于AI的长篇小说仿写工具"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Novel Rewriter Team"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "openai>=1.6.0",
    "httpx>=0.25.0",
    "tiktoken>=0.5.0",
    "jieba>=0.42.1",
    "numpy>=1.26.0",
    "scikit-learn>=1.3.0",
    "rich>=13.7.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.25.0",
]

[project.scripts]
novel-rewriter = "backend.main:main"

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
